host control with active speaker fix
